requirements.txt


Django
django-chartjs
django-import-export
msgpack
django-cors-headers
django-dynamic-raw-id
Pillow
django-file-validator
django-ckeditor
django-nested-admin
django_form_builder
